refactor: sp_plan.py owns the plan contract; sp_attention.py is dispatch-only
SequenceParallelPlan, the boundary-hook interpreter and apply_sequence_parallel
move to sp_plan.py; sp_attention.py keeps only the dispatch-level attention
installer. The two are siblings over sp_ops — the plan holds attention as an
opaque callable, so neither imports the other, and the file layout now matches
the dependency structure (model_backend imports the contract, the family-config
default imports the installer). Pure code motion, no behavior change.
